frontend/requirements.txt
streamlit==1.38.0
requests==2.32.3
PyYAML==6.0.2
google-generativeai
python-dotenv==1.0.1
